Create README.md
I have yet to learn a million things about this world and the field I trust myself to get into.
That's why I am doing my best in honing all my skills to the top-tier, so I can meet amazing people and help the generation in developing the technology we are relying on right now. 

Also I know languages like English, Hindi, Japanese,etc. Still trying to be fluent in all of them. It's a bit challenging but I will definately do it.

So if you find me interesting then please connect with me so we can advance together.